+++ src/main/java/com/memi/lifeos/supplement/repository/SupplementRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SupplementRepository
-		extends JpaRepository<Supplement, Long>, JpaSpecificationExecutor<Supplement> {
+		extends JpaRepository<Supplement, Integer>, JpaSpecificationExecutor<Supplement> {
 }
